refactor: eliminate duplicate inline routes, fix module mount paths (fixes #1895)
- Strip /api prefix from forms, portfolio, notifications, admin route modules
- Enhance notifications module with push persistence, dual auth, preferences
- Add circuit breaker to admin module for membership fetch
- Remove ~540 lines of duplicate inline routes from server/index.js
- Clean up unused imports in server/index.js
- Import loadPersistedPushSubscriptions from notifications module
- server/index.js: 1690 -> 1146 lines
