+++ server/routes/notifications.js
@@ -1,7 +1,8 @@
 /**
  * Notification & Push Subscription Routes
  * Manages browser push subscriptions, notification CRUD,
- * and read-state tracking with rate limiting and admin auth.
+ * read-state tracking, and user preferences with dual auth
+ * (admin or student) and optional database persistence.
  */
 
 import { Router } from 'express';
@@ -10,12 +11,65 @@ import { adminAuthMiddleware } from '../middleware/adminAuthMiddleware.js';
 import { requireStudentAuth } from '../middleware/studentAuthMiddleware.js';
 import { notificationRateLimiter } from '../middleware/rateLimiter.js';
 import notificationsService from '../services/notificationsService.js';
+import { pushSubscriptionsRepository } from '../repositories/pushSubscriptionsRepository.js';
+import { notificationPreferencesRepository } from '../repositories/notificationPreferencesRepository.js';
+import { studentAuthService } from '../services/studentAuthService.js';
 
 const router = Router();
-const adminAuth = adminAuthMiddleware.requireAdmin;
 
-// ── In-memory push subscription store ──────────────────────────────────────
+// ── Constants ───────────────────────────────────────────────────────────────
+const PUSH_PERSISTENCE_ENABLED = Boolean(process.env.DATABASE_URL);
+
+// ── In-memory push subscription store with persistence ──────────────────────
 const pushSubscriptions = new Set();
+
+async function loadPersistedPushSubscriptions() {
+  if (!PUSH_PERSISTENCE_ENABLED) return;
+  try {
+    const rows = await pushSubscriptionsRepository.list({ limit: 10000 });
+    for (const sub of rows) {
+      pushSubscriptions.add(JSON.stringify(sub));
+    }
+    console.log(`Loaded ${rows.length} persisted push subscription(s).`);
+  } catch (err) {
+    console.error('Failed to load persisted push subscriptions:', err.message);
+  }
+}
+
+async function persistPushSubscription(subscription) {
+  if (!PUSH_PERSISTENCE_ENABLED) return;
+  try {
+    await pushSubscriptionsRepository.add(subscription);
+  } catch (err) {
+    console.error('Failed to persist push subscription:', err.message);
+  }
+}
+
+async function removePersistedPushSubscription(subscription) {
+  if (!PUSH_PERSISTENCE_ENABLED) return;
+  try {
+    await pushSubscriptionsRepository.remove(subscription.endpoint);
+  } catch (err) {
+    console.error('Failed to remove persisted push subscription:', err.message);
+  }
+}
+
+export { loadPersistedPushSubscriptions };
+
+// ── Dual auth middleware (admin OR student) ─────────────────────────────────
+function requireNotificationAuth(req, res, next) {
+  adminAuthMiddleware.requireAdmin(req, res, (err) => {
+    if (!err && req.adminSession) {
+      return next();
+    }
+    requireStudentAuth(req, res, (err2) => {
+      if (!err2 && req.studentUser) {
+        return next();
+      }
+      return res.status(401).json({ error: 'Unauthorized: Authentication required' });
+    });
+  });
+}
 
 // ── Push subscription validation middleware ────────────────────────────────
 const validatePushSubscription = [
@@ -52,51 +106,80 @@ const validatePushSubscription = [
   },
 ];
 
+// ── Push Subscription Routes ────────────────────────────────────────────────
+
 /**
- * POST /api/notifications/subscribe — Register a push subscription.
+ * POST /notifications/subscribe — Register a push subscription.
  * Limits total stored subscriptions to 10 000 (FIFO eviction).
  */
-router.post('/api/notifications/subscribe', validatePushSubscription, (req, res) => {
-  try {
-    const { subscription } = req.body;
-    if (subscription) {
-      pushSubscriptions.add(JSON.stringify(subscription));
-      if (pushSubscriptions.size > 10000) {
-        const oldest = pushSubscriptions.values().next().value;
-        pushSubscriptions.delete(oldest);
+router.post(
+  '/notifications/subscribe',
+  adminAuthMiddleware.requireAdmin,
+  notificationRateLimiter,
+  validatePushSubscription,
+  async (req, res) => {
+    try {
+      const { subscription } = req.body;
+      if (subscription) {
+        pushSubscriptions.add(JSON.stringify(subscription));
+        if (pushSubscriptions.size > 10000) {
+          const oldest = pushSubscriptions.values().next().value;
+          pushSubscriptions.delete(oldest);
+        }
+        await persistPushSubscription(subscription);
       }
+      return res.json({ success: true });
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
     }
-    return res.json({ success: true });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
   }
-});
+);
 
 /**
- * POST /api/notifications/unsubscribe — Remove a push subscription.
- */
-router.post('/api/notifications/unsubscribe', validatePushSubscription, (req, res) => {
-  try {
-    const { subscription } = req.body;
-    if (subscription) pushSubscriptions.delete(JSON.stringify(subscription));
-    return res.json({ success: true });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-/**
- * POST /api/notifications/mark-read — Mark a single notification as read.
+ * POST /notifications/unsubscribe — Remove a push subscription.
  */
 router.post(
-  '/api/notifications/mark-read',
-  adminAuth,
+  '/notifications/unsubscribe',
+  adminAuthMiddleware.requireAdmin,
+  notificationRateLimiter,
+  validatePushSubscription,
+  async (req, res) => {
+    try {
+      const { subscription } = req.body;
+      if (subscription) {
+        pushSubscriptions.delete(JSON.stringify(subscription));
+        await removePersistedPushSubscription(subscription);
+      }
+      return res.json({ success: true });
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
+    }
+  }
+);
+
+// ── Notification CRUD Routes (dual auth) ────────────────────────────────────
+
+/**
+ * POST /notifications/mark-read — Mark a single notification as read.
+ */
+router.post(
+  '/notifications/mark-read',
+  requireNotificationAuth,
   notificationRateLimiter,
   async (req, res) => {
     try {
       const { id, userId } = req.body || {};
       if (!id) return res.status(400).json({ error: 'id required' });
-      const uid = userId || 'global';
+      let uid = userId || 'global';
+      if (req.studentUser) {
+        const studentId = req.studentUser.sub || req.studentUser.id;
+        if (userId && userId !== studentId) {
+          return res
+            .status(403)
+            .json({ error: 'Forbidden: Cannot modify other users notifications' });
+        }
+        uid = studentId;
+      }
       const ok = await notificationsService.markAsRead(uid, id);
       return res.json({ success: ok });
     } catch (err) {
@@ -106,16 +189,24 @@ router.post(
 );
 
 /**
- * POST /api/notifications/mark-all-read — Mark all notifications as read.
+ * POST /notifications/mark-all-read — Mark all notifications as read.
  */
 router.post(
-  '/api/notifications/mark-all-read',
-  adminAuth,
+  '/notifications/mark-all-read',
+  requireNotificationAuth,
   notificationRateLimiter,
   async (req, res) => {
     try {
       const { userId } = req.body || {};
-      await notificationsService.markAllAsRead(userId || 'global');
+      let uid = userId || 'global';
+      if (req.studentUser) {
+        const studentId = req.studentUser.sub || req.studentUser.id;
+        if (userId && userId !== studentId) {
+          return res.status(403).json({ error: 'Forbidden' });
+        }
+        uid = studentId;
+      }
+      await notificationsService.markAllAsRead(uid);
       return res.json({ success: true });
     } catch (err) {
       return res.status(500).json({ error: err.message });
@@ -124,71 +215,191 @@ router.post(
 );
 
 /**
- * DELETE /api/notifications/:id — Remove a specific notification by ID.
+ * DELETE /notifications/:id — Remove a specific notification by ID.
  */
-router.delete('/api/notifications/:id', adminAuth, notificationRateLimiter, async (req, res) => {
-  try {
-    const id = req.params.id;
-    const userId = req.query.userId || 'global';
-    const removed = await notificationsService.removeNotification(userId, id);
-    if (!removed) return res.status(404).json({ error: 'Notification not found' });
-    return res.json({ success: true });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-/**
- * DELETE /api/notifications — Clear all notifications for a user.
- */
-router.delete('/api/notifications', adminAuth, notificationRateLimiter, async (req, res) => {
-  try {
-    const userId = req.query.userId || 'global';
-    await notificationsService.clearAll(userId);
-    return res.json({ success: true });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
-  }
-});
-
-/**
- * POST /api/notifications — Create a new notification (admin).
- */
-router.post('/api/notifications', adminAuth, notificationRateLimiter, async (req, res) => {
-  try {
-    const { userId, title, message, type, link } = req.body || {};
-    if (!title || !message) {
-      return res.status(400).json({ error: 'title and message are required' });
+router.delete(
+  '/notifications/:id',
+  requireNotificationAuth,
+  notificationRateLimiter,
+  async (req, res) => {
+    try {
+      const id = req.params.id;
+      let uid = req.query.userId || 'global';
+      if (req.studentUser) {
+        const studentId = req.studentUser.sub || req.studentUser.id;
+        if (req.query.userId && req.query.userId !== studentId) {
+          return res.status(403).json({ error: 'Forbidden' });
+        }
+        uid = studentId;
+      }
+      const removed = await notificationsService.removeNotification(uid, id);
+      if (!removed) return res.status(404).json({ error: 'Notification not found' });
+      return res.json({ success: true });
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
     }
-    const note = await notificationsService.addNotification(userId || 'global', {
-      title,
-      message,
-      type,
-      link,
-    });
-    return res.json({ success: true, notification: note });
-  } catch (err) {
-    return res.status(500).json({ error: err.message });
   }
-});
+);
 
 /**
- * GET /api/notifications — Retrieve notifications for the authenticated user.
- * Requires a valid student token; query userId must match the token subject.
+ * DELETE /notifications — Clear all notifications for a user.
  */
-router.get('/api/notifications', requireStudentAuth, notificationRateLimiter, async (req, res) => {
-  try {
-    const authenticatedUserId = req.studentUser?.sub || req.studentUser?.id;
-    const requestedUserId = req.query.userId || 'global';
+router.delete(
+  '/notifications',
+  requireNotificationAuth,
+  notificationRateLimiter,
+  async (req, res) => {
+    try {
+      let uid = req.query.userId || 'global';
+      if (req.studentUser) {
+        const studentId = req.studentUser.sub || req.studentUser.id;
+        if (req.query.userId && req.query.userId !== studentId) {
+          return res.status(403).json({ error: 'Forbidden' });
+        }
+        uid = studentId;
+      }
+      await notificationsService.clearAll(uid);
+      return res.json({ success: true });
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
+    }
+  }
+);
 
-    if (requestedUserId !== 'global' && requestedUserId !== authenticatedUserId) {
-      return res.status(403).json({ error: 'Forbidden' });
+/**
+ * POST /notifications — Create a new notification (admin).
+ */
+router.post(
+  '/notifications',
+  adminAuthMiddleware.requireAdmin,
+  notificationRateLimiter,
+  async (req, res) => {
+    try {
+      const { userId, title, message, type, link } = req.body || {};
+      if (!title || !message) {
+        return res.status(400).json({ error: 'title and message are required' });
+      }
+      const note = await notificationsService.addNotification(userId || 'global', {
+        title,
+        message,
+        type,
+        link,
+      });
+      return res.json({ success: true, notification: note });
+    } catch (err) {
+      return res.status(500).json({ error: err.message });
+    }
+  }
+);
+
+// ── Notification Retrieval (dual auth) ─────────────────────────────────────
+
+/**
+ * GET /notifications — Retrieve notifications for the authenticated user.
+ * Supports ?userId=, ?offset=N, ?limit=N query params.
+ * Dual auth: admin session or student token.
+ */
+router.get('/notifications', async (req, res) => {
+  try {
+    const userId = req.query.userId || 'global';
+
+    if (userId !== 'global') {
+      let authenticated = false;
+
+      // 1. Try Student Auth
+      let token = null;
+      const authHeader = req.headers.authorization;
+      if (authHeader && authHeader.startsWith('Bearer ')) {
+        token = authHeader.slice(7).trim();
+      }
+      if (!token && req.cookies?.ns_student_token) {
+        token = req.cookies.ns_student_token;
+      }
+      if (token) {
+        const payload = studentAuthService.verifyToken(token);
+        if (payload && (payload.sub === userId || payload.id === userId)) {
+          authenticated = true;
+        }
+      }
+
+      // 2. Try Admin Auth
+      if (!authenticated) {
+        let adminToken = null;
+        if (authHeader && authHeader.startsWith('Bearer ')) {
+          adminToken = authHeader.slice(7).trim();
+        }
+        if (!adminToken && req.cookies?.ns_admin_token) {
+          adminToken = req.cookies.ns_admin_token;
+        }
+        if (adminToken) {
+          const { getAdminSession } = await import('../repositories/adminSessionsRepository.js');
+          const session = await getAdminSession(adminToken);
+          if (session) {
+            authenticated = true;
+          }
+        }
+      }
+
+      if (!authenticated) {
+        return res.status(401).json({ error: 'Unauthorized to view these notifications' });
+      }
     }
 
-    const list = await notificationsService.getNotifications(
-      requestedUserId === 'global' ? 'global' : authenticatedUserId
-    );
+    const offset = parseInt(req.query.offset, 10) || 0;
+    const limit = Math.min(parseInt(req.query.limit, 10) || 100, 500);
+    const list = await notificationsService.getNotifications(userId, offset, limit);
     return res.json({ notifications: list });
+  } catch (err) {
+    return res.status(500).json({ error: err.message });
+  }
+});
+
+// ── Notification Preferences Routes ─────────────────────────────────────────
+
+/**
+ * GET /notifications/preferences — List notification preferences.
+ */
+router.get('/notifications/preferences', async (req, res) => {
+  try {
+    const userId = req.query.userId || 'global';
+    const prefs = await notificationPreferencesRepository.list(userId);
+    return res.json({ preferences: prefs });
+  } catch (err) {
+    return res.status(500).json({ error: err.message });
+  }
+});
+
+/**
+ * PUT /notifications/preferences — Set a single preference.
+ */
+router.put('/notifications/preferences', async (req, res) => {
+  try {
+    const userId = req.body.userId || 'global';
+    const { category, email, push, in_app } = req.body;
+    if (!category) return res.status(400).json({ error: 'category is required' });
+    const pref = await notificationPreferencesRepository.set(userId, category, {
+      email,
+      push,
+      in_app,
+    });
+    return res.json({ preference: pref });
+  } catch (err) {
+    return res.status(500).json({ error: err.message });
+  }
+});
+
+/**
+ * PUT /notifications/preferences/bulk — Set multiple preferences at once.
+ */
+router.put('/notifications/preferences/bulk', async (req, res) => {
+  try {
+    const userId = req.body.userId || 'global';
+    const { preferences } = req.body;
+    if (!Array.isArray(preferences) || !preferences.length) {
+      return res.status(400).json({ error: 'preferences array is required' });
+    }
+    const results = await notificationPreferencesRepository.setBulk(userId, preferences);
+    return res.json({ preferences: results });
   } catch (err) {
     return res.status(500).json({ error: err.message });
   }
